+++ assets/javascripts/kitten/components/buttons/button-tooltip-text.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+export default class ButtonTooltipText extends React.Component {
+  render() {
+    return (
+      <button className="k-Button
+                         k-Button--helium
+                         k-ButtonTooltip--helium">
+              Passez à l'étape 2.
+      </button>
+    )
+  }
+}