+++ spec/dummy/client/entries/app-kitten.js
@@ -9,6 +9,7 @@ import DocLinkBox from 'kitten/components/box/doc-link-box'
 // Buttons
 import ButtonImageWithTextAndBadge from 'kitten/components/buttons/button-image-with-text-and-badge'
 import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
+import ButtonTooltipText from 'kitten/components/buttons/button-tooltip-text'
 
 // Dev
 import { DevGrid } from 'kitten/components/dev/dev-grid'
@@ -37,6 +38,7 @@ import NavList from 'kitten/components/lists/nav-list'
 // Tooltips
 import ReactTooltip from 'react-tooltip'
 import Tooltip from 'kitten/components/tooltips/tooltip'
+import ButtonTooltip from 'kitten/components/tooltips/button-tooltip'
 
 // Tours
 import HeaderTour from 'kitten/components/tours/header-tour'
@@ -65,6 +67,8 @@ import { KarlLoanSimulator,
 
 ReactOnRails.register({
   ButtonImageWithTextAndBadge,
+  ButtonTooltip,
+  ButtonTooltipText,
   ButtonTooltipIcon,
   DevGrid,
   DocLinkBox,